Changed structure a little
