+++ MediHealthProject/app/appStyle.js
@@ -14,26 +14,5 @@ export default StyleSheet.create({
         fontSize: 20,
         fontWeight: '400',
         color: "#FFFFFF",
-    },
-    MedicineButton: {
-        alignItems: 'flex-start',
-        backgroundColor: '#28DA9A',
-        width: 270,
-        height: 60,
-        borderRadius: 8,
-        padding: 15,
-        marginTop:10
-    },
-    MedicineButtonText: {
-        fontSize: 20,
-        fontWeight: '400',
-        color: "#FFFFFF",
-    },
-    medicineInputButton: {
-        alignItems: 'center',
-        height: 50,
-        width: 50,
-        backgroundColor: '#28DA9A',
-        borderRadius: 50
     }
 });
