+++ MediHealthProject/app/containers/AppointmentScreen/appointmentScreen.js
@@ -1,13 +1,13 @@
 import React, { Component } from "react";
 import { View, Text, StyleSheet, Image } from "react-native";
 import { Icon, Button, Container, Content, Left, Right } from "native-base";
-import MyHeader from "../components/header";
+import MyHeader from "../../components/header";
 
-class RehabilitationScreen extends Component {
+class AppointmentScreen extends Component {
     static navigationOptions = {
         drawerIcon: (
             <Image
-                source={require("../assets/images/rehabilitation-icon.png")}
+                source={require("../../assets/images/appointment-icon.png")}
                 style={{ height: 24, width: 24, tintColor: "black" }}
             />
         )
@@ -24,11 +24,11 @@ class RehabilitationScreen extends Component {
                         justifyContent: "center"
                     }}
                 >
-                    <Text>Rehabilitation Screen</Text>
+                    <Text>Appointment Screen</Text>
                 </Content>
             </Container>
         );
     }
 }
 
-export default RehabilitationScreen;
+export default AppointmentScreen;
