+++ MediHealthProject/app/containers/MedicineScreen/medicineScreen.js
@@ -1,14 +1,14 @@
 import React, { Component } from "react";
 import { View, Text, StyleSheet, Image, FlatList, TouchableOpacity } from "react-native";
 import { Icon, Button, Container, Content, Left, Right } from "native-base";
-import MyHeader from "../components/header";
-import styles from '../appStyle'
+import MyHeader from "../../components/header";
+import styles from './appStyle'
 
 class MedicineScreen extends Component {
     static navigationOptions = {
         drawerIcon: (
             <Image
-                source={require("../assets/images/medicine-icon.png")}
+                source={require("../../assets/images/medicine-icon.png")}
                 style={{ height: 24, width: 24, tintColor: "black" }}
             />
         )
@@ -20,12 +20,13 @@ class MedicineScreen extends Component {
                 <MyHeader nav={this.props.navigation} />
                 <Content
                     contentContainerStyle={{
-                        flex: 1,
-                        alignItems: "center",
-                        justifyContent: "center"
+                        flex: 1
+                        //alignItems: "center",
+                        //width: "100%"
+                        //justifyContent: "center"
                     }}
                 >
-                    <Text>Medicine Screen</Text>
+                
                     <FlatList
                         data={[
                             {key: 'Panadol'}, 
@@ -41,16 +42,21 @@ class MedicineScreen extends Component {
                             {key: 'j'}]
                         }
                         renderItem={({item}) => 
-                            <TouchableOpacity
-                                title={item.key}
-                                style={styles.MedicineButton}
-                                accessibilityLabel={item.key}
-                                onPress={() => this.props.navigation.navigate('Biomarkers')}
-                            >
-                                <Text style={styles.MedicineButtonText}>{item.key}</Text>
-                            </TouchableOpacity>}
+                            <View style={styles.MedicineButtonContainer}>
+                                <View style={styles.MedicineButtonPadding}></View>
+                                <TouchableOpacity
+                                    title={item.key}
+                                    style={styles.MedicineButton}
+                                    accessibilityLabel={item.key}
+                                    onPress={() => this.props.navigation.navigate('Biomarker')}
+                                >
+                                    <Text style={styles.MedicineButtonText}>{item.key}</Text>
+                                </TouchableOpacity>
+                                <View style={styles.MedicineButtonPadding}></View>
+                            </View>
+}
                     />
-                    <Text>Medicine Screen</Text>
+                    
                     <TouchableOpacity
                         title="MedicineInput"
                         style={styles.medicineInputButton}
