+++ MediHealthProject/app/containers/RehabilitationScreen/rehabilitationScreen.js
@@ -1,13 +1,13 @@
 import React, { Component } from "react";
 import { View, Text, StyleSheet, Image } from "react-native";
 import { Icon, Button, Container, Content, Left, Right } from "native-base";
-import MyHeader from "../components/header";
+import MyHeader from "../../components/header";
 
-class AppointmentScreen extends Component {
+class RehabilitationScreen extends Component {
     static navigationOptions = {
         drawerIcon: (
             <Image
-                source={require("../assets/images/appointment-icon.png")}
+                source={require("../../assets/images/rehabilitation-icon.png")}
                 style={{ height: 24, width: 24, tintColor: "black" }}
             />
         )
@@ -24,11 +24,11 @@ class AppointmentScreen extends Component {
                         justifyContent: "center"
                     }}
                 >
-                    <Text>Appointment Screen</Text>
+                    <Text>Rehabilitation Screen</Text>
                 </Content>
             </Container>
         );
     }
 }
 
-export default AppointmentScreen;
+export default RehabilitationScreen;
